before eval

+++ shell.c
@@ -6,7 +6,7 @@
 /*   By: cdupuis <cdupuis@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/10 13:32:14 by cdupuis           #+#    #+#             */
-/*   Updated: 2023/11/28 11:03:44 by cdupuis          ###   ########.fr       */
+/*   Updated: 2023/11/29 14:56:15 by cdupuis          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,46 +37,11 @@ void	launch_cmd(t_cmd *cmd, t_data *data, char **tokens)
 	exit(EXIT_SUCCESS);
 }
 
-int	shell_execute(char **tokens, t_data *data)
-{
-	t_cmd	*cmd;
-	int		status;
-	int		builtins;
-
-	if (!tokens)
-		return (0);
-	cmd = data->cmd;
-	builtins = 0;
-	status = 0;
-	if (tokens == NULL)
-		return (1);
-	if (tokens[0] == NULL)
-		return (1);
-	if (!cmd->next && !cmd->fds)
-		builtins = launch_builtins(cmd, data, tokens);
-	while (cmd && builtins == 0)
-	{
-		data->pid = fork();
-		if (data->pid == -1)
-			return (0);
-		else if (data->pid == 0)
-		{
-			launch_cmd(cmd, data, tokens);
-		}
-		cmd = cmd->next;
-	}
-	close_pipes(data->cmd, NULL);
-	waitpid(data->pid, &status, WUNTRACED);
-	return (1);
-}
-
 void	shell_loop(t_data *data)
 {
 	char	*line;
 	char	**tokens;
-	int		status;
 
-	status = 0;
 	tokens = NULL;
 	while (1)
 	{
@@ -87,13 +52,16 @@ void	shell_loop(t_data *data)
 		{
 			tokens = shell_split_tokens(data, line);
 			pipe_count(data, tokens);
-			fill_cmd(tokens, &data->cmd);
-			status = shell_execute(tokens, data);
-			if (data->cmd->fds && data->cmd->fds->type == 3)
-				unlink(data->cmd->fds->name);
+			if (data->err == 0)
+			{
+				fill_cmd(tokens, &data->cmd);
+				shell_execute(tokens, data);
+				if (data->cmd->fds && data->cmd->fds->type == 3)
+					unlink(data->cmd->fds->name);
+			}
 			free_str(tokens);
 			clean_cmd(&data->cmd);
+			data->err = 0;
 		}
-		(void)status;
 	}
 }
